requirements.txt
pytesseract==0.3.2
selenium==3.141.0
Pillow==7.1.2